tool_function.py
import os
from dotenv import load_dotenv
from agents import Agent, Runner, AsyncOpenAI, OpenAIChatCompletionsModel, function_tool
from agents.run import RunConfig
import asyncio
import datetime


# Load the environment variables from the .env file
load_dotenv()

gemini_api_key = os.getenv("GEMINI_API_KEY")

# Check if the API key is present; if not, raise an error
if not gemini_api_key:
    raise ValueError("GEMINI_API_KEY is not set. Please ensure it is defined in your .env file.")

#Reference: https://ai.google.dev/gemini-api/docs/openai
external_client = AsyncOpenAI(
    api_key=gemini_api_key,
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/",
)

model = OpenAIChatCompletionsModel(
    model="gemini-2.0-flash",
    openai_client=external_client
)

config = RunConfig(
    model=model,
    model_provider=external_client,
    tracing_disabled=True
)

@function_tool
def get_weather(city: str) -> str:
    return f"The weather in {city} is sunny"

@function_tool
def get_date_time():
    return datetime.date.today()

agent = Agent(
    name="Haiku agent",
    # instructions="Always respond in haiku form",
    instructions="You are an helpfull assistant",
    model=model,
    tools=[get_weather, get_date_time],
)

async def main():
    while True:
        user_input = input("You: ")
        result = await Runner.run(agent, 
                                  user_input, 
                              run_config=config
                              )
        if user_input.lower() == "quit":
            break
        # print(result.new_items[2])
        print(result.new_items[2].raw_item.content[0].text)
      



if __name__ == "__main__":
    asyncio.run(main())